build: use docker compose override for dev
Using a `compose-override.yml` allows me to easily use volume mounts in
dev, while keeping a separate `compose.yml` that is meant for
production.

It seems that the Compose config. files need to be named `compose.yml`
instead of `docker-compose.yml` in order for the override to work.

This also adds the `--watch` command to Deno in order to reload the
server when a file in the dependency tree has been updatd.